requirements.txt
# UI
streamlit==1.30.0

# LangChain core
langchain==0.1.16
langchain-community==0.0.34
langchain-text-splitters==0.0.1

# FREE LLM Provider
langchain-groq==0.1.3

# Vector Database (Python 3.11 compatible)
faiss-cpu==1.9.0.post1

# PDF Loading
pypdf==4.0.1

# FREE Embeddings
sentence-transformers==2.2.2

# Environment Variables
python-dotenv==1.0.1

# Utilities
tqdm==4.66.1
